[feat] controller, service 폴더 생성
