+++ src/public/css/reservation/equipment/step1.css
@@ -1,0 +1,383 @@
+.wrapper {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+form {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+}
+.first_form {
+  padding-left: 25px;
+  margin-top: 15px;
+}
+.step-sub {
+  font-size: 18px;
+  font-weight: bold;
+
+  color: #8d091d;
+}
+.step_box {
+  margin: 17px 0;
+  width: 100%;
+  height: 44px;
+  display: flex;
+  flex-direction: row;
+}
+.step {
+  width: 50%;
+  color: white;
+  font-family: sans-serif;
+  font-weight: 600;
+  font-size: 18px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
+}
+#step1 {
+  background-color: #8d091d;
+}
+#step2 {
+  background-color: #cbcbcb;
+}
+.standard {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  justify-content: flex-end;
+  padding-bottom: 13px;
+  padding-right: 4px;
+}
+#num {
+  font-size: 13px;
+  margin: 0;
+  color: #707070;
+}
+#time {
+  font-size: 13px;
+  margin: 0;
+  color: #8d091d;
+}
+.red_menu_box {
+  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
+  background-color: #8d091d;
+  width: 100%;
+  height: 50px;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0px 35px;
+  border-bottom: #740314 0.5px solid;
+}
+.red_menu_text {
+  color: white;
+  margin: 0;
+}
+.white_menu_box {
+  background-color: white;
+  display: flex;
+  width: 100%;
+  height: 46px;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  box-shadow: 0 3px 6px 0 rgba(0, 0, 0, 0.16);
+  background-color: #eeeeee;
+  padding: 0px 35px;
+}
+.white_menu_text {
+  color: black;
+  margin: 0;
+}
+.cam_box {
+  padding-top: 2px;
+  width: 100%;
+  max-height: 2000px;
+  transition: height 0.8s ease-in;
+  -webkit-transition: height 0.8s ease-in;
+  /* animation-name: DropDownOn;
+  animation-duration: 0.8s;
+  animation-fill-mode: forwards; */
+}
+/* @keyframes DropDownOn {
+  from {opacity: 0}
+  to {opacity: 1}
+}
+@keyframes DropDownOff {
+  0% {opacity: 1}
+  100% {opacity: 0}
+} */
+.cam_menu {
+  margin: 0 15px;
+  /* background-color: blue; */
+  display: flex;
+  flex-direction: column;
+  align-items: left;
+}
+.cam_check {
+  padding: 10px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  /* background-color: cadetblue; */
+}
+.cam_check p {
+  margin: 0 10px;
+}
+.graph {
+  border-top: 0.3px solid black;
+  margin-bottom: 10px;
+}
+.chart {
+  display: flex;
+  flex-direction: row;
+}
+.redbar {
+  background-color: #8d091d;
+  width: 2px;
+  height: 100px;
+  padding-right: 2px;
+}
+.time-block {
+  width: 18px;
+  height: 8px;
+  border-radius: 2px;
+  border: solid 0.5px #707070;
+  background-color: #dfdfdf;
+}
+.time-hour {
+  display: flex;
+  flex-direction: row-reverse;
+  justify-content: flex-end;
+  align-items: flex-end;
+  font-size: 11px;
+  padding: 0.8px;
+}
+.day {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  margin-left: 5px;
+  margin-right: 5px;
+}
+.hour_num {
+  font-size: 11px;
+  /* background-color: aqua;  */
+  word-spacing: 7px;
+  margin-top: 4px;
+  line-height: 0.8;
+  margin-bottom: 0;
+  letter-spacing: -0.7px;
+}
+.day_divide {
+  width: 1px;
+  border-left: dashed 1px #707070;
+}
+.stack {
+  display: flex;
+  flex-direction: row;
+  margin-left: 6px;
+}
+.bar_body {
+  display: flex;
+  flex-direction: row;
+  width: 85%;
+  white-space: nowrap;
+  overflow-x: scroll;
+}
+.dotline {
+  height: 1px;
+  margin-top: 2px;
+  margin-bottom: 2px;
+  top: 27px;
+  position: relative;
+  z-index: 10;
+  border-top: dotted 1px #707070;
+}
+#blank {
+  height: 5px;
+  opacity: 0;
+}
+.blank {
+  height: 30px;
+  opacity: 0;
+}
+.grayline {
+  height: 1px;
+  color: #707070;
+}
+.timeline p {
+  font-size: 14px;
+  margin-top: 2px;
+}
+.flex {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+}
+#circle {
+  border-radius: 50%;
+  border: 2px black solid;
+}
+
+.borrow_list {
+  margin: 0 40px;
+  list-style-type: circle;
+}
+
+.button_box {
+  width: 100%;
+  height: 50px;
+  margin-bottom: 50px;
+  margin-top: 30px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+button {
+  width: 155px;
+  height: 50px;
+  border-radius: 25px;
+  border: none;
+  box-shadow: 3px 3px 10px 0 rgba(0, 0, 0, 0.16);
+  background-image: linear-gradient(to bottom, #aa1a30, #8d091d);
+  font-size: 16px;
+  font-weight: bold;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: 1.5;
+  letter-spacing: normal;
+  text-align: center;
+  color: #ffffff;
+  margin-bottom: 0px;
+}
+.rev_time {
+  margin: 5px;
+}
+.set_time {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin: 20px;
+}
+#fromDate {
+  width: 90%;
+  margin-bottom: 3px;
+}
+#returnDate {
+  width: 100%;
+  margin-bottom: 3px;
+}
+.hihi {
+  padding-left: 20px;
+}
+
+.mystyle {
+  max-height: 0;
+  transition: height 0.8s ease-out;
+  -webkit-transition: height 0.8s ease-out;
+  overflow: hidden;
+
+  /* animation-name: DropDownOff;
+  animation-duration: 0.5s;
+  animation-fill-mode: forwards; */
+}
+#arrow {
+  font-size: 14px;
+}
+
+
+#camera {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/unclicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+#camera:checked {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/clicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+
+#etc {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/unclicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+#etc:checked {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/clicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+
+#카메라보조장치 {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/unclicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+#etc:checked {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/clicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+
+.edit {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/unclicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+.edit:checked {
+  width: 20px;
+  height: 20px;
+  appearance: none;
+  background-image: url("/static/images/clicked.png");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: auto 100%;
+}
+.countSelect {
+  font-size: 14px;
+  font-weight: 500;
+  font-stretch: normal;
+  font-style: normal;
+  line-height: 1.43;
+  letter-spacing: normal;
+  text-align: left;
+  color: #4f4f4f;
+}
+.none{
+  display: none;
+}
